feat: move all changes from SimpleTextDisplayer to new NativeTextDisplayer
