+++ lib/text/text_utils.js
@@ -6,6 +6,7 @@
 
 goog.provide('shaka.text.Utils');
 
+goog.require('goog.asserts');
 goog.require('shaka.log');
 goog.require('shaka.text.Cue');
 goog.require('shaka.text.CueRegion');
@@ -313,5 +314,115 @@ shaka.text.Utils = class {
     shaka.text.Cue.parseCuePayload(cue);
 
     return cue;
+  }
+
+  /**
+   * @param {!TextTrack} textTrack
+   * @param {!Array<shaka.text.Cue>} cues
+   */
+  static appendCuesToTextTrack(textTrack, cues) {
+    const flattenedCues = shaka.text.Utils.getCuesToFlatten(cues);
+
+    // Convert cues.
+    const textTrackCues = [];
+    const cuesInTextTrack = textTrack.cues ?
+      Array.from(textTrack.cues) : [];
+
+    for (const inCue of flattenedCues) {
+      // When a VTT cue spans a segment boundary, the cue will be duplicated
+      // into two segments.
+      // To avoid displaying duplicate cues, if the current textTrack cues
+      // list already contains the cue, skip it.
+      const containsCue = cuesInTextTrack.some((cueInTextTrack) => {
+        if (cueInTextTrack.startTime == inCue.startTime &&
+          cueInTextTrack.endTime == inCue.endTime &&
+          cueInTextTrack.text == inCue.payload) {
+          return true;
+        }
+        return false;
+      });
+
+      if (!containsCue && inCue.payload) {
+        const cue =
+          shaka.text.Utils.mapShakaCueToNativeCue(inCue);
+        if (cue) {
+          textTrackCues.push(cue);
+        }
+      }
+    }
+
+    // Sort the cues based on start/end times.  Make a copy of the array so
+    // we can get the index in the original ordering.  Out of order cues are
+    // rejected by Edge.  See https://bit.ly/2K9VX3s
+    const sortedCues = textTrackCues.slice().sort((a, b) => {
+      if (a.startTime != b.startTime) {
+        return a.startTime - b.startTime;
+      } else if (a.endTime != b.endTime) {
+        return a.endTime - b.startTime;
+      } else {
+        // The browser will display cues with identical time ranges from the
+        // bottom up.  Reversing the order of equal cues means the first one
+        // parsed will be at the top, as you would expect.
+        // See https://github.com/shaka-project/shaka-player/issues/848 for
+        // more info.
+        // However, this ordering behavior is part of VTTCue's "line" field.
+        // Some platforms don't have a real VTTCue and use a polyfill instead.
+        // When VTTCue is polyfilled or does not support "line", we should not
+        // reverse the order.  This occurs on legacy Edge.
+        // eslint-disable-next-line no-restricted-syntax
+        if ('line' in VTTCue.prototype) {
+          // Native VTTCue
+          return textTrackCues.indexOf(b) - textTrackCues.indexOf(a);
+        } else {
+          // Polyfilled VTTCue
+          return textTrackCues.indexOf(a) - textTrackCues.indexOf(b);
+        }
+      }
+    });
+
+    for (const cue of sortedCues) {
+      textTrack.addCue(cue);
+    }
+  }
+
+
+  /**
+   * Iterate over all the cues in a text track and remove all those for which
+   * |predicate(cue)| returns true.
+   *
+   * @param {!TextTrack} track
+   * @param {function(!TextTrackCue):boolean} predicate
+   */
+  static removeCuesFromTextTrack(texTrack, predicate) {
+    // Since |track.cues| can be null if |track.mode| is "disabled", force it
+    // to something other than "disabled".
+    //
+    // If the texTrack is already showing, then we should keep it as showing.
+    // But if it something else, we will use hidden so that we don't "flash"
+    // cues on the screen.
+    let disabled = false;
+
+    if (texTrack.mode === 'disabled') {
+      disabled = true;
+      texTrack.mode = 'hidden';
+      goog.asserts.assert(
+        texTrack.cues,
+        'Cues should be accessible when mode is set to "hidden".',
+      );
+    }
+
+    let i = 0;
+    while (i < texTrack.cues.length) {
+      const cue = texTrack.cues[i];
+      if (predicate(cue)) {
+        texTrack.removeCue(cue);
+      } else {
+        i++;
+      }
+    }
+
+    if (disabled) {
+      texTrack.mode = 'disabled';
+    }
   }
 };
